# web-rally/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rally Tascas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/rally/v1/teams/{team_id}/badges": {
      "get": {
        "tags": [
          "Badges"
        ],
        "summary": "List Team Badges",
        "description": "All badges a single team holds, newest first.",
        "operationId": "list_team_badges",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamBadgeRead"
                  },
                  "title": "Response List Team Badges"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rally/v1/teams/{team_id}/badge-showcase": {
      "get": {
        "tags": [
          "Badges"
        ],
        "summary": "Team Badge Showcase",
        "description": "The full badge board for a team: active catalogue + what it has earned.\n\nOne request drives the locked/earned grid. ``earned`` collapses to the\nearliest award per badge code (a team may hold a code for several scopes).\nWhen the feature is disabled the board is empty so a direct API hit\nreveals nothing.",
        "operationId": "team_badge_showcase",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadgeShowcaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rally/v1/badge-definitions": {
      "get": {
        "tags": [
          "Badge Admin"
        ],
        "summary": "List Badge Definitions",
        "operationId": "list_badge_definitions",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BadgeDefinitionResponse"
                  },
                  "type": "array",
                  "title": "Response List Badge Definitions"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Badge Admin"
        ],
        "summary": "Create Badge Definition",
        "operationId": "create_badge_definition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BadgeDefinitionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadgeDefinitionResponse"
                }
              }
            }
          },
          "409": {
            "description": "Badge code already exists"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/rally/v1/badge-definitions/{id}": {
      "put": {
        "tags": [
          "Badge Admin"
        ],
        "summary": "Update Badge Definition",
        "operationId": "update_badge_definition",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BadgeDefinitionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadgeDefinitionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Badge definition not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Badge Admin"
        ],
        "summary": "Delete Badge Definition",
        "operationId": "delete_badge_definition",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Badge definition not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rally/v1/badge-definitions/{id}/icon": {
      "post": {
        "tags": [
          "Badge Admin"
        ],
        "summary": "Upload Badge Icon",
        "operationId": "upload_badge_icon",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_badge_icon"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadgeDefinitionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Badge definition not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rally/v1/badges/award": {
      "post": {
        "tags": [
          "Badge Admin"
        ],
        "summary": "Manual Award Badge",
        "operationId": "manual_award_badge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualBadgeAwardCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamBadgeRead"
                }
              }
            }
          },
          "404": {
            "description": "Badge definition not found"
          },
          "400": {
            "description": "Badge is inactive"
          },
          "409": {
            "description": "Team already holds this badge"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/rally/v1/badges/{badge_id}": {
      "delete": {
        "tags": [
          "Badge Admin"
        ],
        "summary": "Revoke Badge",
        "operationId": "revoke_badge",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "badge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Badge Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Badge not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BadgeDefinitionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "is_auto": {
            "type": "boolean",
            "title": "Is Auto",
            "default": false
          },
          "color": {
            "type": "string",
            "title": "Color",
            "default": "#8b5cf6"
          },
          "glyph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glyph"
          },
          "trigger_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Type"
          },
          "criteria": {
            "additionalProperties": true,
            "type": "object",
            "title": "Criteria",
            "default": {}
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "name",
          "code"
        ],
        "title": "BadgeDefinitionCreate"
      },
      "BadgeDefinitionLite": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "glyph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glyph"
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "color"
        ],
        "title": "BadgeDefinitionLite",
        "description": "Display fields of a badge definition, for the team showcase."
      },
      "BadgeDefinitionResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "is_auto": {
            "type": "boolean",
            "title": "Is Auto",
            "default": false
          },
          "color": {
            "type": "string",
            "title": "Color",
            "default": "#8b5cf6"
          },
          "glyph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glyph"
          },
          "trigger_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Type"
          },
          "criteria": {
            "additionalProperties": true,
            "type": "object",
            "title": "Criteria",
            "default": {}
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "code"
        ],
        "title": "BadgeDefinitionResponse"
      },
      "BadgeDefinitionUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "is_auto": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Auto"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "glyph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Glyph"
          },
          "trigger_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Type"
          },
          "criteria": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criteria"
          }
        },
        "type": "object",
        "title": "BadgeDefinitionUpdate"
      },
      "BadgeShowcaseResponse": {
        "properties": {
          "definitions": {
            "items": {
              "$ref": "#/components/schemas/BadgeDefinitionLite"
            },
            "type": "array",
            "title": "Definitions"
          },
          "earned": {
            "items": {
              "$ref": "#/components/schemas/EarnedBadge"
            },
            "type": "array",
            "title": "Earned"
          }
        },
        "type": "object",
        "required": [
          "definitions",
          "earned"
        ],
        "title": "BadgeShowcaseResponse",
        "description": "Everything the team badge board needs in one request.\n\n``definitions`` is the full active catalogue (drives the locked+earned grid);\n``earned`` is this team's awards, joined to the catalogue by ``code``."
      },
...
      "Body_upload_badge_icon": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_upload_badge_icon"
      },
...
      "EarnedBadge": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "awarded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Awarded At"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "code",
          "awarded_at"
        ],
        "title": "EarnedBadge",
        "description": "A badge a team has earned, keyed by definition code."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualBadgeAwardCreate": {
        "properties": {
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "badge_code": {
            "type": "string",
            "title": "Badge Code"
          },
          "activity_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Id"
          },
          "checkpoint_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Id"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "badge_code"
        ],
        "title": "ManualBadgeAwardCreate"
      },
...
      "TeamBadgeRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "badge_type": {
            "type": "string",
            "title": "Badge Type"
          },
          "activity_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Id"
          },
          "checkpoint_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Id"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta",
            "default": {}
          },
          "awarded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Awarded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "badge_type",
          "awarded_at"
        ],
        "title": "TeamBadgeRead",
        "description": "A single badge a team holds."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}